+++ CHANGELOG.md
@@ -30,7 +30,10 @@
 * Update the display of list of Subject headings. Refs UIIN-815.
 * Add enumeration and chronology data to the item table on the instance record. Refs UIIN-757.
 * Add filters for staffSuppress and discoverySuppress. Completes UIIN-289.
+* Display empty elements with the dashes. Refs UIIN-826.
 * Filter item records by material type. Part of UIIN-777.
+* Filter item records by item status. Part of UIIN-771.
+* Allow for searching title with words out of order. Part of UIIN-602.
 
 ## [1.12.1](https://github.com/folio-org/ui-inventory/tree/v1.12.1) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.12.0...v1.12.1)